feat: Implement converting University of Auckland GPA to honours-level
And vice versa.  The intent of this feature is to be able to get whether
one gets First Class Honours or Second Class Honours from their GPA.  If
they get Second Class Honours, their division (First Division or Second
Division) is also determined.

The reverse is also true, that is, by specifying an honours level (and
division, if wanted), the GPA range is output.